+++ tests/MediaTest/MediaTest.iOS/ViewController.cs
@@ -46,7 +46,11 @@ namespace MediaTest.iOS
 
             PickPhoto.TouchUpInside += async (sender, args) =>
             {
-                var test = await CrossMedia.Current.PickPhotoAsync();
+                var test = await CrossMedia.Current.PickPhotoAsync(
+                    new Plugin.Media.Abstractions.PickMediaOptions
+                    {
+                        PhotoSize = SizeSwitch.On ? Plugin.Media.Abstractions.PhotoSize.Medium : Plugin.Media.Abstractions.PhotoSize.Full
+                    });
                 if (test == null)
                     return;